\name{callsleavingqueue-methods}
\docType{methods}
\alias{callsleavingqueue-methods}
\title{Methods for Function \code{callsleavingqueue}}
\description{
function to calculate proportion of calls resulting in customer hanging up
}
\section{Methods}{
\describe{
\item{\code{signature(z = "mc")}}{
z parameter means the object of class "mc"
}
}}

