\name{exphit-methods}
\docType{methods}
\alias{exphit-methods}
\title{Methods for Function \code{exphit}}
\description{
Methods for claculating expected hitting time of transforming from state i to state j
}
\section{Methods}{
\describe{
\item{\code{signature(z = "mc")}}{
z parameter means the object of class "mc"
}
\item{\code{signature(i = "ANY")}}{
state i
}
\item{\code{signature(j = "ANY")}}{
state j
}
}}
\keyword{Expecting hitting time}

