\name{findpi-methods}
\docType{methods}
\alias{findpi-methods}
\title{ Methods for Function \code{findpi} }
\description{
   Methods for claculating pi
}
\section{Methods}{
\describe{
\item{\code{signature(z = "mc")}}{
%%  x parameter means the object of class "mc"
}

\item{\code{signature(k = "ANY")}}{
%%  k parameter means the rows number of the matrix
}
}}
\keyword{calculate pi}
\keyword{transition matrix}

