\name{findpicontin-methods}
\docType{methods}
\alias{findpicontin-methods}
\title{Methods for Function \code{findpicontin}}
\description{
 function to find pi vector
}
\section{Methods}{
\describe{
\item{\code{signature(z = "mc")}}{
z parameter means the object of class "mc"
}
}}
\keyword{ find pi vector}

