\name{findq-methods}
\docType{methods}
\alias{findq-methods}
\title{Methods for Function \code{findq}}
\description{
 Methods for claculating the Q matrix
}
\section{Methods}{
\describe{

\item{\code{signature(z = "mc")}}{
z parameter means the object of class "mc"
}
}}
\keyword{methods}
\keyword{Q matrix}
