\name{rowsb-methods}
\docType{methods}
\alias{rowsb-methods}
\title{Methods for Function \code{rowsb}}
\description{
 Methods for calculating the number of rows and colums for the Q matrix based on all combinations of i+j<= b
}
\section{Methods}{
\describe{
\item{\code{signature(z = "mc")}}{
z parameter means the object of class "mc"
}
\item{\code{signature(b = "ANY")}}{
maximum calls in the system
}
}}
\keyword{the number of rows and colums for the Q matrix}
